refactor: Separate CoordinationStrategy to Modules
CoordinationStrategy now serves as a wrapper around the individual
modules, with a AWSStrategy calling functions through AWS in mind.

The functionality of the app is split into modules. The modules may use
shared utilities such as persistance and llm classes, but should not
depend on other modules. The modules are prepared to be exposed as
individual aws lambdas.